backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.1
html5lib==1.1
httpx==0.24.1
backoff==2.2.1
python-docx==1.1.0
python-multipart
